change directDebitGuarantee to render in diferent places with different text copy prefixs
